Big Screen Completed
The pre big text screen now saves the details that you put in for next time you launch the app. Big text now displays the users name and whether they are alone or not and to take them to the police station. Side function of the app is now complete.

Project: Hanover